ai-server/requirements.txt
fastapi
uvicorn
pillow
torch
transformers
faiss-cpu
python-multipart
requests
pymongo
numpy